feat: advertise realtime voice provider voices in talk.catalog
Realtime voice providers could not tell Talk clients which voices they
support, so browser and native Talk UIs fell back to a hard-coded
OpenAI-only voice list that is wrong for every other realtime provider.

Add an optional `voices?: readonly string[]` to RealtimeVoiceProviderPlugin
and project it into talk.catalog realtime entries. The projection lives in
core (a shared talk-catalog.ts module split out of talk.ts), so plugins
cannot inject the field themselves. Speech, transcription, and realtime
entries now share one helper that advertises models/aliases/voices only
when non-empty, which also fixes transcription and realtime entries
silently dropping provider-declared models.

The field advertises each provider's canonical set of known selectable
voice ids. OpenAI advertises the ten ids its config validation accepts;
the default gpt-realtime-2.1 accepts all of them and marin/cedar are its
recommended voices. Google Live advertises the shared GOOGLE_TTS_VOICES
inventory, which its native-audio models accept and the runtime forwards
unvalidated.
